Refactor make box mesh code
If the make_box_mesh.h file is included into multiple compilation units
all of the in-header functions get compiled into each unit. Many of the
functions are declared `static` so they don't collide. However, one is not
and that actually causes linker-errors.

Rather than just declaring the single offending function static/inline,
this moves the implementation into a .cc file to improve overall
compilation time. There are three functions that only exist to serve the
implementation of the mesh generation. Their declarations have been kept
in the header file because they are directly tested in make_box_mesh_test.

Also, the function comments were changed from /** */ to /* */ reflecting
their non-doxygen, internal:: status.

Otherwise, this introduces no new code and doesn't change the text of any
comments.
